.backup_pre_refactor/20251031_141651/backend_full/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "my-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "websockets",
    "requests",
    "apscheduler",
    "pandas>=2.3.3",
    "ccxt>=4.0.0",
    "hyperliquid-python-sdk>=0.20.0",
    "python-dotenv>=1.2.1",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py"]

